Move find_err to ErrorPath.line_col
